+++ 100_design_common/010_revisitplan/helper_scripts/support_fieldwork_organization.R
@@ -132,73 +132,13 @@ scheme_moco_ps_stratum_targetpanel_spsamples %>%
 ## Sampling unit geometries --------------------------------------
 
 # obtaining geometries of the sampling units themselves:
-# - for lentic types, the result is in object 'watersurface_spsamples_sf' below
+# - for lentic types, the result is in object 'stratum_grts_spsamples_lentic_sf'
+#   from the REP
 # - for lotic types, see code in
 #   inbo/n2khab-mne-monitoring/010_aq_piezometer_positioning, but then do use
 #   the REP RData file used here
 # - for type 7220 (springs) as a whole, see code provided below
 # - for terrestrial types, these are cells; see code provided below
-
-
-
-# geometries of the spatial sampling units of lentic types (polygons)
-# /////////////////////////////////////////////////////////////////////////
-
-# link between spatial sampling units of lentic types and watersurface polygons
-stratum_grts_polygon_spsamples_lentic <-
-  scheme_moco_ps_stratum_sppost_spsamples %>%
-  filter(str_detect(stratum, "^2190_a|^31")) %>%
-  unnest(sp_poststr_samples) %>%
-  select(-sample_status) %>%
-  # provision for potential local replacements by including grts_address_final
-  # (currently not different for these strata)
-  add_assessment_data() %>%
-  distinct(stratum, grts_address_final) %>%
-  inner_join(
-    units_non_cell_n2khab_grts %>%
-      filter(sample_support_code == "watersurface") %>%
-      select(-sample_support_code),
-    join_by(grts_address_final == grts_address),
-    relationship = "many-to-one",
-    unmatched = c("error", "drop")
-  ) %>%
-  rename(polygon_id = unit_id)
-
-# extract all polygons from watersurfaces_hab that belong to the spatial samples
-# for lentic types:
-wsh_polygons_spsamples <-
-  read_watersurfaces_hab(version = versions_required["watersurfaces_hab"]) %>%
-  pluck("watersurfaces_polygons") %>%
-  select(polygon_id) %>%
-  semi_join(stratum_grts_polygon_spsamples_lentic, join_by(polygon_id))
-
-# extract all polygons from the watersurfaces data source that belong to the
-# spatial samples for lentic types but that were not covered by
-# watersurfaces_hab:
-ws_extra_polygons_spsamples <-
-  read_watersurfaces(
-    version = versions_required["watersurfaces"],
-    fix_geom = TRUE
-  ) %>%
-  select(polygon_id) %>%
-  semi_join(stratum_grts_polygon_spsamples_lentic, join_by(polygon_id)) %>%
-  # keeping only the unique extra polygons relative to watersurfaces_hab
-  anti_join(
-    st_drop_geometry(wsh_polygons_spsamples),
-    join_by(polygon_id)
-  )
-
-# add the geometry to the spatial sampling units:
-stratum_grts_spsamples_lentic_sf <-
-  rbind(wsh_polygons_spsamples, ws_extra_polygons_spsamples) %>%
-  inner_join(
-    stratum_grts_polygon_spsamples_lentic,
-    join_by(polygon_id),
-    relationship = "one-to-many",
-    unmatched = "error"
-  ) %>%
-  relocate(stratum, grts_address_final) %>%
-  arrange(stratum, grts_address_final)
 
 
 
@@ -456,7 +396,6 @@ units_cell_polygon %>%
   mutate(overlapped_cell_area = st_area(.)) %>%
   st_drop_geometry() %>%
   rename(grts_address_overlapped_cell = grts_address_final)
-
 
 
 
@@ -1971,6 +1910,19 @@ if (FALSE) {
       layer = "orthophotoscreening_shortterm_lentictypes_WSPOLYGONS",
       delete_layer = TRUE
     )
+
+  # generating & writing layer of legacy watersample points for lentic types,
+  # restricted to the locations in the shortterm calendar
+  legacy_watersamplepoints_spslocs_lentic %>%
+    semi_join(
+      fieldwork_shortterm_prioritization_watersurfaces,
+      join_by(grts_address_final)
+    ) %>%
+    write_sf(
+      gpkg_path,
+      layer = "legacywatsamppoints_shortterm_lentictypes_POINTS",
+      delete_layer = TRUE
+    )
 }
 
 
@@ -1984,7 +1936,6 @@ objects <- tibble(
     "stratum_schemepstargetpanel_spsamples",
     "schemepstargetpanel_spsamples_terr",
     "vbi_overlaps",
-    "stratum_grts_polygon_spsamples_lentic",
     "stratum_grts_spsamples_lentic_sf",
     "units_7220",
     "units_cell_cellcenter",